Append first shortcuts
